requirements.txt
openai>=1.6.0
streamlit>=1.29.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
tqdm>=4.66.0
plotly>=5.18.0
scikit-learn>=1.3.0
